splitting out main routines from api
