+++ bin/jobsub
@@ -14,12 +14,13 @@ PREFIX = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(os.path.join(PREFIX, "lib"))
 
 # pylint: disable=wrong-import-position, ungrouped-imports
-from jobsub_api import (
+from mains import (
     jobsub_submit_args,
     jobsub_cmd_args,
     jobsub_fetchlog_args,
     jobsub_cmd_parser,
     jobsub_fetchlog_parser,
+    VERBOSE,
 )
 
 from get_parser import (
@@ -27,8 +28,6 @@ from get_parser import (
     get_condor_epilog,
 )
 
-
-VERBOSE = 0
 
 if __name__ == "__main__":
     try:
@@ -58,9 +57,7 @@ if __name__ == "__main__":
         args.func(args, passthru)
 
     except Exception as e:  # pylint: disable=broad-except
-        # if VERBOSE:
-        # pylint: disable=using-constant-test
-        if True:
+        if VERBOSE:
             raise
         sys.stderr.write(f"\n\nError: {e.__class__.__name__}: {str(e)}\n\n")
         sys.exit(1)
