+++ bin/jobsub_submit
@@ -19,7 +19,6 @@
 
 """ submit command for jobsub layer over condor """
 # pylint: disable=wrong-import-position,wrong-import-order,import-error
-import hashlib
 import os
 import os.path
 import sys
@@ -39,14 +38,13 @@ PREFIX = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 # this means we don't need fancy ups dependencies..
 #
 sys.path.append(os.path.join(PREFIX, "lib"))
-from jobsub_api import jobsub_submit_main
-
+from mains import jobsub_submit_main, VERBOSE
 
 if __name__ == "__main__":
     try:
         jobsub_submit_main()
     except Exception as e:  # pylint: disable=broad-except
-        if verbose:
+        if VERBOSE:
             raise
         sys.stderr.write(f"\n\nError: {e.__class__.__name__}: {str(e)}\n\n")
         sys.exit(1)
